+++ aioradio/long_running_jobs.py
@@ -24,9 +24,8 @@ from aioradio.redis import Redis
 
 @dataclass
 class LongRunningJobs:
-    """Worker that continually pulls from queue (either SQS or Redis list
-    implemented like queue), running a job using request parameters conveyed in
-    the message.
+    """Worker that continually pulls from SQS queue, running a job using
+    request parameters conveyed in the message.
 
     Also has functions to send messages to the queue, and to check if
     job is complete.
@@ -46,12 +45,10 @@ class LongRunningJobs:
     # Flexibility to define one to many jobs, ex: {"job1_name": (async_func1, 30), "job2_name": (async_func2, 60)}
     # First item of tuple must be an async function that runs your long running job, and the second item the
     # job timeout, which should be a value ~3x or more above the max time a job finishes corresponding to the
-    # visibility_timeout if queue_service = 'sqs' or message re-entry into not-started queue if queue_service = 'redis'.
+    # visibility_timeout.
     jobs: Dict[str, Tuple[Any, float]] = field(default_factory=dict)
 
-    # choose between sqs or redis
-    queue_service: str = 'sqs'
-    # if using sqs than define the queue name and aws region
+    # define the queue name and aws region
     sqs_queue: str = None
     sqs_region: str = None
 
@@ -64,16 +61,13 @@ class LongRunningJobs:
     httpx_client: httpx.AsyncClient = httpx.AsyncClient()
 
     def __post_init__(self):
-        self.queue_service = self.queue_service.lower()
-        if self.queue_service not in ['sqs', 'redis']:
-            raise ValueError("queue_service must be either 'sqs' or 'redis'.")
-
         if self.fakeredis:
             self.cache = Redis(fake=True)
         else:
             self.cache = Redis(config={'redis_primary_endpoint': self.redis_host, 'encoding': 'utf-8'})
 
         self.job_names = set(self.jobs.keys())
+
         for job_name, job_info in self.jobs.items():
             if len(job_info) != 2:
                 raise ValueError('Job info should be provided as a tuple, ex. (job_function, job_timeout)')
@@ -136,11 +130,8 @@ class LongRunningJobs:
         result = {}
         try:
             msg = orjson.dumps(items).decode()
-            if self.queue_service == 'sqs':
-                entries = [{'Id': str(uuid4()), 'MessageBody': msg, 'MessageGroupId': self.host_uuid}]
-                await sqs.send_messages(queue=self.sqs_queue, region=self.sqs_region, entries=entries)
-            else:
-                self.cache.pool.rpush(f'{job_name}-not-started', msg)
+            entries = [{'Id': str(uuid4()), 'MessageBody': msg, 'MessageGroupId': self.host_uuid}]
+            await sqs.send_messages(queue=self.sqs_queue, region=self.sqs_region, entries=entries)
 
             await self.cache.hmset(
                 key=identifier,
@@ -158,24 +149,13 @@ class LongRunningJobs:
 
         while not self.stop:
             try:
-                # run job the majority of the time pulling up to 10 messages to process
-                for _ in range(10):
-                    if self.queue_service == 'sqs':
-                        await self.__sqs_pull_messages_and_run_jobs__()
-                    else:
-                        for job_name in self.job_names:
-                            await self.__redis_pull_messages_and_run_jobs__(job_name)
-
-                # verify processing only a fraction of the time
-                if self.queue_service == 'redis':
-                    for job_name in self.job_names:
-                        await self.__verify_processing__(job_name)
+                await self.__sqs_pull_messages_and_run_jobs__()
             except asyncio.CancelledError:
                 print(traceback.format_exc())
                 break
             except Exception:
                 print(traceback.format_exc())
-                await asyncio.sleep(30)
+                await asyncio.sleep(5)
 
         self.stop = False
 
@@ -233,83 +213,13 @@ class LongRunningJobs:
             total = round(time() - float(msg[0]['Attributes']['SentTimestamp'])/1000, 3)
             print(f"Async college match processing time for UUID {body['uuid']} took {total} seconds")
 
-    async def __redis_pull_messages_and_run_jobs__(self, job_name):
-        """Pull messages one at a time and run job.
-
-        Args:
-            job_name (str): Name of job corresponding to a key in self.jobs
-
-        Raises:
-            IOError: Redis access failed
-        """
-
-        # in the future convert lpop to lmove and also look into integrating with async aioredis
-        msg = self.cache.pool.lpop(f'{job_name}-not-started')
-        if not msg:
-            await asyncio.sleep(0.1)
-        else:
-            body = orjson.loads(msg)
-            key = body['cache_key']
-
-            callback_url = body['params'].get('callback_url', '')
-            body['params'].pop('callback_url', None)
-
-            # Add start time and push msg to <self.name>-in-process
-            body['start_time'] = time()
-            self.cache.pool.rpush(f'{job_name}-in-process', orjson.dumps(body).decode())
-
-            data = None if key is None else await self.cache.get(key)
-            if data is None:
-                # No results found in cache so run the job
-                data = await self.jobs[job_name][0](body['params'])
-
-                # Set the cached parameter based key with results
-                if key is not None and not await self.cache.set(key, data, expire=self.expire_cached_result):
-                    raise IOError(f"Setting cache string failed for cache_key: {key}")
-
-            # Send results via POST request if necessary
-            if callback_url:
-                create_task(self.httpx_client.post(callback_url, json={'results': data, 'uuid': body['uuid']}, timeout=30))
-
-            # Update the hashed UUID with processing results
-            await self.cache.hmset(
-                key=body['uuid'],
-                items={**body, **{'results': data, 'job_done': True}},
-                expire=self.expire_job_data
-            )
-
-    async def __verify_processing__(self, job_name):
-        """Verify processing completed fixing issues related to app crashing or
-        scaling down servers when using queue_service = 'redis'.
-
-        Args:
-            job_name (str): Name of job corresponding to a key in self.jobs
-        """
-
-        timeout = self.jobs[job_name][1]
-        for _ in range(10):
-            msg = self.cache.pool.lpop(f'{job_name}-in-process')
-            if not msg:
-                break
-
-            body = orjson.loads(msg)
-            job_done = await self.cache.hget(key=body['uuid'], field='job_done')
-            if job_done is None:
-                pass    # if the cache is expired then we can typically ignore doing anything
-            elif not job_done:
-                if (time() - body['start_time']) > timeout:
-                    print(f'{job_name} with uuid: {body["uuid"]} failed after {timeout} seconds. \
-                            Pushing msg back to {job_name}-not-started.')
-                    self.cache.pool.rpush(f'{job_name}-not-started', msg)
-                else:
-                    self.cache.pool.rpush(f'{job_name}-in-process', msg)
-
     @staticmethod
     async def build_cache_key(params: Dict[str, Any], separator='|') -> str:
-        """Build a cache key from a dictionary object.  Concatenate and
+        """Build a cache key from a dictionary object.
+
+        Concatenate and
         normalize key-values from an unnested dict, taking care of sorting the
         keys and each of their values (if a list).
-
         Args:
             params (Dict[str, Any]): dict object to use to build cache key
             separator (str, optional): character to use as a separator in the cache key. Defaults to '|'.